+++ src/components/Board.svelte
@@ -9,6 +9,7 @@
     import type { IRenderOptions } from "@abstractplay/renderer";
     import { onMount } from "svelte";
     import ColorPicker from "svelte-awesome-color-picker";
+    import { generateField as genField } from "@/lib/modular";
 
     const boardTypes = new Map<SupportedBoards, string>([
         [
@@ -128,14 +129,19 @@
     let whichWidth: "abs" | "minmax" | undefined;
     let canBlock = false;
     let canAlternate = false;
+    let canHalf = false;
     let symmetryLocked = true;
     let invertOrientation = false;
     let canInvertOrientation = false;
     let startDiamonds = false;
     let selectedOptions: ValidOption[] = [];
+    let selectedHalf: "full"|"top"|"bottom" = "full";
+    let numModules = 7;
     const initVars = () => {
         canBlock = false;
         canAlternate = false;
+        canHalf = false;
+        selectedHalf = "full";
         if ($state.board.style === undefined) {
             whichWidth = undefined;
         } else if (
@@ -185,6 +191,7 @@
             if ($state.board.style.startsWith("hex-of")) {
                 canBlock = true;
                 canAlternate = true;
+                canHalf = true;
             }
         } else {
             // go board here
@@ -199,6 +206,16 @@
 
     $: if (selectedOptions !== undefined) {
         $state.options = selectedOptions;
+        $state = $state;
+    }
+
+    $: if (selectedHalf !== undefined) {
+        if (selectedHalf === "full") {
+            delete $state.board.half;
+        } else {
+            $state.board.alternatingSymmetry = false;
+            $state.board.half = selectedHalf;
+        }
         $state = $state;
     }
 
@@ -360,6 +377,11 @@
         $state = $state;
         return false;
     };
+
+    const generateField = () => {
+        $state.board = genField(numModules) as (BoardBasic & { style: SupportedBoards | "modular-hex"; });
+        $state = $state;
+    }
 
     let previewDiv: HTMLDivElement;
     onMount(() => {
@@ -544,6 +566,22 @@
                 </label>
             </div>
         {/if}
+        {#if canHalf}
+        <div class="control">
+            <label class="radio">
+                <input type="radio" name="half" value="full" bind:group="{selectedHalf}" />
+                Full
+            </label>
+            <label class="radio">
+              <input type="radio" name="half" value="top" bind:group="{selectedHalf}" />
+              Top half
+            </label>
+            <label class="radio">
+              <input type="radio" name="half" value="bottom" bind:group="{selectedHalf}" />
+              Bottom half
+            </label>
+          </div>
+        {/if}
         {#if canInvertOrientation}
             <div class="control">
                 <label class="checkbox">
@@ -605,6 +643,19 @@
                 }}">Clear blocks</button
             >
         {/if}
+        <hr />
+        <div class="content">
+            <p>You can generate a randomized hex field composed of modules (hexhex 2s) arranged to always touch another at two points. Select the number of modules and then the `Generate field` button.</p>
+        </div>
+        <div class="field padTop">
+            <label class="label" for="numModules">Number of modules</label>
+            <div class="control">
+                <input class="input" name="numModules" type="number" min="1" max="30" bind:value="{numModules}" />
+            </div>
+            <div class="control">
+                <button class="button apButton" on:click="{generateField}">Generate field</button>
+            </div>
+        </div>
         <hr />
         <div class="content">
             <p>