+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.6.1] - 2025-01-13
+
+### Added
+
+-   Added support for half rendering of `hex-of-*` boards.
+-   Added the randomized modular hex field option.
+
 ## [1.6.0] - 2024-11-13
 
 ### Added